Added the `ActorRef` struct
`ActorRef` serves as a strongly-typed wrapper on top of `ActorCell` to automate a lot of the type mapping that needs to go one regarding getting the appropriate message and state types.

It also exposes a suite of functions related to rpcs, timers, and the registry which utilize the `TActor` which is encoded into the struct, removing the need to explicitely state it in the call.