Added subscribe-able Output ports
This change adds support for another struct called an `OutputPort` which is a publish-subscribe forwarder from the output from one actor to the input to another.

It manages the lifecycle of the subscription automatically and lets you have an actor emit messages without actively knowing who's listening (and at what point).

Resolves: #4