Adding PG groups
Erlang's process groups are named groups of collections of actors. They provide
easy to use reference material for actors which may have "dispatching" purposes,
in that you may want to send to 1, all, a region, etc.

This change also includes making state mutable. We've removed a lot of extra tasks
and migrated to using the future's crate `catch_unwind` which saves a lot of state
synchronization code laying around.