Cleanup trait definitions
This change renames `ActorHandler` to `Actor` and `Actor` to `ActorRuntime`. This is helpful because 90% of what people will work with is just the trait definition, which `Actor` is more logical than the "handler" naming scheme.

Additionally we can avoid existing breakages by re-defining on the `Actor{Handler}` trait the `spawn` and `spawn_linked` calls such that anyone already calling `Actor::spawn` doesn't have to change any syntax.