Adding support for "instant" spawning (#48)
* Adding support for "instant" spawning

This PR adds support for spawning actors "instantly" in that we don't wait on the `pre_start` routine. This is helpful if you want to immediately have access to the `ActorRef<TActor>` so you can begin sending messages to it without blocking on the `pre_start` routine.

**However** if the `pre_start` fails for any reason, it won't be captured unless the specific join handle is handled manually to monitor the spawning error. Documentation updated to reflect this.

New tests covering instant spawning added.

Additionally we're adding a small auto-implementation for types which implement `BytesSerializable` such that they are automatically implementing `ractor::Message` and can be network-sent. This makes most of the built-in numeric types, `()`, `String`s, and vectors of such able to immediately used in a cluster context.

* version bump