Add support for "stop and wait" and "kill and wait"
This change adds support for stopping or killing an actor while waiting for the shutdown routine to complete. This is often useful to be assured an actor's shutdown routine has completed before trying to

1. respawn the same actor (potential name clashes/etc)
2. Exiting the process, which may "leak" memory if a thread is still alive without first waiting for proper actor shutdown.

This is often easier than waiting on the `JoinHandle` which is needed to be threaded and shutdown may not be executed by the same process which created/spawned the actor and recorded the `JoinHandle`.

Ideally the `JoinHandle`s wouldn't be necessary, and for ergonomics this would be enough.