Code re-organization for better flow.
This should be a functional no-op, but has been on my backlog for a bit. I'm re-organizing unnecessary sub-folders and whatnot into a flatter structure that makes logically more sense.

All re-exported types are the same, but the original definition of many things have changed so most users should se no difference, but it may cause some advanced cases to need to adjust imports/usages.