Improve NodeSession TCP performance
Each of NodeSession's message is encoded with a 2 bytes length
prefix followed by the actual object data. Originally when writing
the message to TcpStream we write them separately, however the
TcpStream is not buffered, so the first 2 bytes write may be pushed
to the underlying network stack as a tiny packet, then the data
is pushed as one or more packets.

The first tiny packet might trigger the pathological interaction
between Nagle's Algorithm and Delayed Acknowledgement in TCP, causing
from 40ms to 200ms delay of the actual payload delivery. This can
be observed from the average 40ms latency between two nodes on the
same host conneting via the loopback interface.

Combine the length prefix and payload data to one write_all call
to avoid this latency anomaly. Since the message is already
length prefixed, also skip the length delimiter encoding in proto.

Signed-off-by: Kan-Ru Chen <kanru@kanru.info>