Add ability for a factory's internal state to be updated dynamically (#333)
This allows someone to say reload configeration settings, and internally tear-down and rebuild the factory without fully shutting down and restarting.

Some settings are not dynamically settable without tearing down and rebuilding the factory, unless we want to introduce more strong-types to the message type. This includes, unfortunately, the worker builder, queue, and router.

For settings on those two, there are two choices. (1) tear down and rebuild the factory or (2) thread in dynamic settings to your router and/or queue protocols such that they can internally change their functionality.