thread-local actors (#349)
This PR adds supports for a new actor trait, ThreadLocalActor, which doesn't require the actor now it's internal state to be Send.

This however does require that it be thread-local and be pinned to execution on a single thread.

Additionally, due to the actor itself not being Send, we require that the actor handler implement Default so that it can be constructed on the local thread, only when needed.

This should additionally enforce best-practice, as &self in actors shouldn't hold variables, rather should be delgated to Arguments and State